refactor helpers in interpreter/robot + tests
