+++ droidlet/interpreter/tests/test_interpret_attribute.py
